+++ pkg/node/http_handlers.go
@@ -3,12 +3,10 @@ package node
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"io"
 	"log/slog"
 	"net/http"
 	"os"
-	"strconv"
 	"sync"
 	"time"
 )
@@ -75,26 +73,6 @@ func (s *Node) Forward(w http.ResponseWriter, req *http.Request, owner string) {
 
 }
 
-func readBody(req *http.Request) ([]byte, error) {
-	b, err := io.ReadAll(req.Body)
-	if err != nil && err.Error() != "EOF" {
-		return nil, err
-	}
-	return b, nil
-}
-
-func parseTTL(req *http.Request) (time.Duration, error) {
-	ttlStr := req.URL.Query().Get("ttl")
-	if ttlStr == "" {
-		return 0, nil
-	}
-	sec, err := strconv.Atoi(ttlStr)
-	if err != nil {
-		return 0, fmt.Errorf("invalid ttl")
-	}
-	return time.Duration(sec) * time.Second, nil
-}
-
 // put adds a key/value pair, writing to all replicas.
 func (n *Node) Put(w http.ResponseWriter, req *http.Request) {
 	key := req.URL.Path[len("/kv/"):]
@@ -151,27 +129,6 @@ func (n *Node) Put(w http.ResponseWriter, req *http.Request) {
 	w.WriteHeader(http.StatusNoContent)
 }
 
-// putReplica writes the key/value pair to the local store (called by the primary).
-func (n *Node) PutReplica(w http.ResponseWriter, req *http.Request) {
-	key := req.URL.Path[len("/replica/"):]
-	slog.Info("[Writing Replica]", "url", req.URL.Path, "self id", n.addr)
-
-	body, err := readBody(req)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	ttl, err := parseTTL(req)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	n.kv.Put(key, body, ttl)
-	w.WriteHeader(http.StatusNoContent)
-}
-
 // get returns the value for a key
 func (n *Node) Get(w http.ResponseWriter, req *http.Request) {
 	key := req.URL.Path[len("/kv/"):]
@@ -195,13 +152,6 @@ func (n *Node) Get(w http.ResponseWriter, req *http.Request) {
 	}
 	w.Header().Set("Content-Type", "application/octet-stream")
 	w.Write(val)
-}
-
-func (n *Node) DelReplica(w http.ResponseWriter, req *http.Request) {
-	key := req.URL.Path[len("/replica/"):]
-	slog.Info("[Deleting Replica]", "url", req.URL.Path, "self id", n.addr)
-	n.kv.Delete(key)
-	w.WriteHeader(http.StatusNoContent)
 }
 
 // del removes a key
